xapian_enquire_set_query(): Always call to Xapian
Don't optimise away the call to Xapian if the query object is the one
already set as qlen may be different.  Xapian::Enquire::set_query() is
very cheap as it just takes and stores a references to the query object
and stores an integer.

(No regression test because I can't see how to write one using only
the parts of Xapian's C++ API which xapian-glib currently wraps, so
this may well currently be a latent bug - I spotted it from looking
at the code.)